wordle.py: fix word scoring
Previously the word scoring did not match the scoring in the online
wordle game. This change fixes the calc_score method.

In the online version, each character in the secret can only be used
once to 'light up' a tile. If there are repeated letters in the guess,
letters in the guess are matched first if they are in the same position.

More concretely, here are some examples of the previous / new
behavior:

| secret | guess | previous scoring method | new scoring method|
|---|---|---|---|
| ylems | neeld | 12321 | 11321 |
| abbey | abaca | 33212 | 33111 |
| banal | union | 12112 | 12111 |
| banal | alloy | 22211 | 22111 |
| banal | annal | 22333 | 22333 |

There is a [Reddit
post](https://www.reddit.com/r/wordle/comments/ry49ne/illustration_of_what_happens_when_your_guess_has/)
with more details.